+++ packages/shared/src/Picasso/Picasso.tsx
@@ -204,6 +204,7 @@ const Viewport = () => {
       )
       setWarned(true)
     }
+
     return null
   }